requirements.txt
scrapy=2.8.0
scrapy-playwright=1.40.0
fastapi=0.103.0
uvicorn=0.20.0
psycopg2=2.9.9 
hydra-core
